python-server/requirements.txt
packaging==23.1
paho-mqtt==1.6.1
pip-check==2.8.1
terminaltables==3.1.10
